# lib/api-zod/src/generated/types/attributeVolumeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atmosphere Storytelling System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttributeName } from "./attributeName";

export interface AttributeVolumeResult {
  attribute: AttributeName;
  volume: number;
}